moving factory to the controller
